+++ src/models/vector.py
